ported to bot.js function into a static on Source to return an object with quote and source, wrote passing test and renamed test file
