+++ lib/models/Source.js
@@ -1,4 +1,6 @@
 const mongoose = require('mongoose');
+const Markov = require('markov-chains-text').default;
+const createNewSentence = require('../createNewSentence');
 
 const schema = new mongoose.Schema({
   fullName: {
@@ -17,6 +19,28 @@ schema.virtual('texts', {
   localField: '_id',
   foreignField: 'source'
 });
+
+schema.statics.returnQuoteObject = function(fullName) {
+  return this
+    .findOne({ fullName: fullName })
+    .populate('texts')
+    .then(res => {
+      let textStr = '';
+      //console log for testing only
+      console.log(res);
+      res.texts.forEach(text => {
+        textStr += ' ' + text;
+      });
+      return [textStr, res.fullName];
+    })
+    .then(text => {
+      const markov = new Markov(text[0]);
+      return {
+        quote: createNewSentence(markov),
+        source: text[1]
+      };
+    });
+};
 
 
 module.exports = mongoose.model('Source', schema);
