+++ lib/models/source.test.js
@@ -1,9 +1,9 @@
-const Source = require('../models/Source');
+const Source = require('./Source');
 
 describe('Source model', () => {
+  
   it('has a required full name', () => {
     const source = new Source({});
-
     const { errors } = source.validateSync();
     expect(errors.fullName.message).toEqual('Path `fullName` is required.');
   });
